source_code_final/requirements.txt
Pillow==9.4.
numpy==1.24.2
seaborn==0.13.
dlib==19.24.0
matplotlib==3.8.3
scikit-learn==1.4.1.post1
face-recognition==1.3.0